starken_mimansa_apps/starken_mimansa_apps/requirements.txt
asgiref==3.5.2
astroid==2.13.3
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
cx-Oracle==8.3.0
defusedxml==0.7.1
dill==0.3.6
Django==4.1.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.10.1
fonttools==4.38.0
fpdf2==2.6.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.16.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.3.0
platformdirs==2.6.2
psycopg2==2.9.5
pycparser==2.21
pylint==2.15.10
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.5
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
urwid==2.1.2
wrapt==1.14.1